requirements.txt
numpy==1.19.5
beautifulsoup4==4.9.3
bokeh==2.2.3
multiprocess==0.70.11.1
nltk==3.5
pandas==1.1.5
pandas_datareader==0.9.0
pathos==0.2.7
praw==7.1.1
psaw==0.1.0
python_dateutil==2.8.1
pytz==2021.1
wget==3.2
